backend/requirements.txt
flask
flask-cors
groq
google-generativeai
PyPDF2
python-dotenv
supabase
resend
